+++ example/simple_app/test/flutter_test_config.dart
@@ -4,12 +4,19 @@ import 'package:adaptive_test/adaptive_test.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter_test/flutter_test.dart';
 
+final pixel5TwoButton = pixel5ThreeButton.copyWith(
+  name: 'pixel_5_two_button',
+  systemNavBar: const SystemNavBarData.twoButton(),
+);
+
 final defaultDeviceConfigs = {
   iPhone8,
   iPhone16,
   iPadPro,
   desktop,
   pixel9,
+  pixel5ThreeButton,
+  pixel5TwoButton
 };
 
 Future<void> testExecutable(FutureOr<void> Function() testMain) async {