+++ lib/adaptive_test.dart
@@ -1,10 +1,15 @@
 export 'src/adaptive/adaptive_test.dart';
 export 'src/adaptive/devices_data.dart';
 export 'src/adaptive/widgets/adaptive_wrapper.dart';
-export 'src/adaptive/window_configuration.dart';
-export 'src/adaptive/window_size.dart';
+export 'src/adaptive/window_configuration_tester.dart';
+export 'src/adaptive/window_config.dart';
 export 'src/configuration.dart';
 export 'src/helpers/await_images.dart';
 export 'src/helpers/fonts_loader.dart';
 export 'src/helpers/goldens_difference.dart';
 export 'src/helpers/skip_test_extension.dart';
+export 'src/adaptive/window_config_data/window_config_data.dart';
+export 'src/adaptive/window_config_data/system_nav_bar_data.dart'
+    show SystemNavBarData;
+export 'src/adaptive/window_config_data/punch_hole_data.dart';
+export 'src/adaptive/window_config_data/dynamic_island_data.dart';