+++ lib/src/adaptive/widgets/adaptive_wrapper.dart
@@ -1,7 +1,8 @@
 import 'package:adaptive_test/src/adaptive/widgets/layers/hardware_layer.dart';
 import 'package:adaptive_test/src/adaptive/widgets/layers/keyboard_layer.dart';
-import 'package:adaptive_test/src/adaptive/widgets/layers/system_layer.dart';
-import 'package:adaptive_test/src/adaptive/window_size.dart';
+import 'package:adaptive_test/src/adaptive/widgets/layers/system_nav_bar_layer.dart';
+import 'package:adaptive_test/src/adaptive/window_config_data/window_config_data.dart';
+import 'package:adaptive_test/src/adaptive/window_config.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
@@ -9,7 +10,7 @@ import 'package:flutter_test/flutter_test.dart';
 ///
 /// See also:
 /// * [HardwareLayer]
-/// * [SystemLayer]
+/// * [SystemNavBarLayer]
 /// * [KeyboardLayer]
 class AdaptiveWrapper extends StatelessWidget {
   const AdaptiveWrapper({
@@ -27,7 +28,7 @@ class AdaptiveWrapper extends StatelessWidget {
     return WindowConfig(
       windowConfig: windowConfig,
       child: HardwareLayer(
-        child: SystemLayer(
+        child: SystemNavBarLayer(
           child: KeyboardLayer(
             tester: tester,
             child: child,