+++ lib/src/adaptive/widgets/layers/hardware_layer.dart
@@ -1,4 +1,4 @@
-import 'package:adaptive_test/src/adaptive/window_size.dart';
+import 'package:adaptive_test/src/adaptive/window_config.dart';
 import 'package:flutter/material.dart';
 
 /// Widget wrapper that paint hardware elements over a child.