+++ lib/src/configuration.dart
@@ -1,4 +1,5 @@
-import 'package:adaptive_test/src/adaptive/window_size.dart';
+import 'package:adaptive_test/src/adaptive/window_config_data/window_config_data.dart';
+import 'package:adaptive_test/src/adaptive/window_config.dart';
 import 'package:flutter/material.dart';
 
 /// Singleton class that configures global variables for the test.