+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## 0.10.0
+
+* feat: Add support for 2-button and 3-button navigation for Android devices.
+  ```dart
+  WindowConfigData(
+    systemNavBar = SystemNavBarData.threeButton(),
+  )
+  // or
+  WindowConfigData(
+    systemNavBar = SystemNavBarData.twoButton(),
+  )
+  ```
+* **BREAKING**: refactor: replaced `homeIndicator` by `systemNavBar` in `WindowConfigData` for the home gesture indicator:
+  ```diff
+    WindowConfigData(
+  -  homeIndicator = HomeIndicatorData(bottomPadding, size)
+  +  systemNavBar = SystemNavBarData.gestureIndicator(bottomPadding, size)
+    )
+  ```
+
 ## 0.9.1
 
 * feat: await image now works with ShapeDecopration